Trying to solve the issue with the KAN MNIST
Set the update_grid to only run the updating function on layers except
the first one. My assumption is that since MNIST images are made of a
lot of 0s, the updating code fails to configure the grid size. Not
updating the grid based on the input dataset is not an issue, since we
already know that all the pixel values are between 0 and 1.